Rename app module to application
