+++ authark/application/services/token_service.py
@@ -1,7 +1,7 @@
 import json
 from typing import Dict
 from abc import ABC, abstractmethod
-from authark.app.models.token import Token
+from authark.application.models.token import Token
 
 
 class TokenService(ABC):
