+++ authark/infra/config/registry.py
@@ -1,9 +1,10 @@
 from collections import UserDict
 from typing import Any, Dict
-from authark.app.models.user import User
+from authark.application.models.user import User
 
-from authark.app.coordinators.auth_coordinator import AuthCoordinator
-from authark.app.repositories.user_repository import MemoryUserRepository
+from authark.application.coordinators.auth_coordinator import AuthCoordinator
+from authark.application.repositories.user_repository import (
+    MemoryUserRepository)
 from authark.infra.crypto.pyjwt_token_service import PyJWTTokenService
