+++ authark/infra/crypto/pyjwt_token_service.py
@@ -1,7 +1,7 @@
 from typing import Dict
 import jwt
-from authark.app.services.token_service import TokenService
-from authark.app.models.token import Token
+from authark.application.services.token_service import TokenService
+from authark.application.models.token import Token
 
 
 class PyJWTTokenService(TokenService):
